[expo-image][android] Moved ResizeMode to enums folder & rename
